Base Interceptors: Moved to own files
