+++ server/src/response/response.decorator.ts
@@ -2,9 +2,11 @@
 import { Type, UseInterceptors } from "@nestjs/common";
 import { ApiPaginatedResponse } from "../pagination/paginated-response.decorator";
 import { ApiOkResponse } from "@nestjs/swagger";
-import ResponseBuilderInterceptor, { ArrayResponseBuilderInterceptor, PaginatedResponseBuilderInterceptor } from "./response.interceptor";
 import ResponseType from "./response-type.enum";
 import Constructor from "src/utils/constructor";
+import ArrayResponseBuilderInterceptor from "./interceptors/array-response.interceptor";
+import PaginatedResponseBuilderInterceptor from "./interceptors/page-response.interceptor";
+import ResponseBuilderInterceptor from "./interceptors/response.interceptor";
 
 type ResponseDecoratorParam<ToType extends Type<unknown>, FromType = unknown> = {
 	handler: Constructor<ResponseBuilderInterceptor<FromType, ToType>>
